ingestor/requirements.txt
click==8.1.3
fastf1==3.0.7
Flask==2.3.2
google-cloud-storage==2.10.0
gunicorn==20.1.0
loguru==0.7.0
motor==3.3.0
pymongo==4.5.0
pytz==2023.3
tqdm==4.66.1
